pyproject.toml
[project]
name = "portable-agent-runner"
version = "0.1.0"
description = "Portable agent runtime with model backends and benchmark adapters"
authors = [{name = "Codex"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.2",
  "httpx>=0.26.0",
  "mini-swe-agent>=0.1.0",
  "pydantic>=2.6.0",
  "pytest>=8.0.0",
  "python-dotenv>=1.0.1",
  "datasets>=2.19.0",
  "ijson>=3.2.3",
]

[project.scripts]
agent = "scripts.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N"]
ignore = []

[tool.pyright]
typeCheckingMode = "basic"
